+++ Engine/Source/Render/RenderUI/Public/MagicUIContext.h
@@ -37,7 +37,7 @@ namespace RenderUI
         void IntSlider(const std::string& label, int* value, int min, int max);
         void FloatSlider(const std::string& label, float* value, float min, float max);
         bool DragFloat(const std::string& label, float* value, float speed = 1.0f, float min = 0.0f, float max = 0.0f,
-            const char* fmt = "%.3f", float power = 1.0f);
+            const char* fmt = "%.3f", int power = 1);
         bool DragInt(const std::string& label, int* value, float speed = 1.0f, int min = 0, int max = 0,
             const char* fmt = "%d");
